src/requirements.txt
streamlit>=1.25.0
requests
geopy
groq
streamlit-mic-recorder
openai-whisper
torch
transformers
gtts
altair
folium
streamlit-folium
icalendar
bleach
pydantic>=2.0.0
streamlit-shadcn-ui